+++ AoE2ScenarioParser/objects/data_objects/condition.py
@@ -1,19 +1,22 @@
 from __future__ import annotations
 
 from enum import IntEnum
-from typing import Union
+from typing import Union, Any
 
 from AoE2ScenarioParser.datasets import conditions
+from AoE2ScenarioParser.datasets.conditions import ConditionId
 from AoE2ScenarioParser.helper.helper import raise_if_not_int_subclass, validate_coords
 from AoE2ScenarioParser.helper.string_manipulations import add_tabs
 from AoE2ScenarioParser.objects.aoe2_object import AoE2Object
 from AoE2ScenarioParser.objects.support.attr_presentation import transform_condition_attr_value
+from AoE2ScenarioParser.objects.support.trigger_object import TriggerComponent
 from AoE2ScenarioParser.sections.retrievers.retriever_object_link import RetrieverObjectLink
 from AoE2ScenarioParser.sections.retrievers.support import Support
 
 
-class Condition(AoE2Object):
+class Condition(AoE2Object, TriggerComponent):
     """Object for handling a condition."""
+    hidden_attribute = 'condition_type'
 
     _link_list = [
         RetrieverObjectLink("condition_type", "Triggers",
@@ -55,9 +58,11 @@ class Condition(AoE2Object):
         RetrieverObjectLink("timer_id", "Triggers",
                             "trigger_data[__index__].condition_data[__index__].timer_id", Support(since=1.46)),
         RetrieverObjectLink("victory_timer_type", "Triggers",
-                            "trigger_data[__index__].condition_data[__index__].victory_timer_type", Support(since=1.46)),
+                            "trigger_data[__index__].condition_data[__index__].victory_timer_type",
+                            Support(since=1.46)),
         RetrieverObjectLink("include_changeable_weapon_objects", "Triggers",
-                            "trigger_data[__index__].condition_data[__index__].include_changeable_weapon_objects", Support(since=1.46)),
+                            "trigger_data[__index__].condition_data[__index__].include_changeable_weapon_objects",
+                            Support(since=1.46)),
         RetrieverObjectLink("xs_function", "Triggers",
                             "trigger_data[__index__].condition_data[__index__].xs_function", Support(since=1.40)),
     ]
@@ -123,6 +128,13 @@ class Condition(AoE2Object):
 
         super().__init__(**kwargs)
 
+    def should_be_displayed(self, attr: str, val: Any) -> bool:
+        # Include the only exception to the -1 == invalid rule
+        if self.condition_type == ConditionId.DIFFICULTY_LEVEL and attr == 'quantity' and val == -1:
+            return True
+
+        return super().should_be_displayed(attr, val)
+
     def get_content_as_string(self, include_effect_definition=False) -> str:
         if self.condition_type not in conditions.attributes:
             attributes_list = conditions.empty_attributes
@@ -132,7 +144,7 @@ class Condition(AoE2Object):
         return_string = ""
         for attribute in attributes_list:
             val = getattr(self, attribute)
-            if attribute == "condition_type" or val in [[], [-1], [''], "", " ", -1]:
+            if not self.should_be_displayed(attribute, val):
                 continue
 
             value_string = transform_condition_attr_value(self.condition_type, attribute, val, self._host_uuid)